v0.5
Completely overhauled meshing to allow for updating the chunkmesh block by block instead of remeshing the entire chunk on a block update

New meshing is much slower to mesh than the old meshing function, but updating the mesh is about infinitely quicker, future version will combine the two
so that chunks closer to the player are able to be updated very quickly and chunks further away are able to be generate much more quickly but are unable to be updated as they never need to be anyways.

Began work on lighting. Basic flood fill algorithm, barely functional right now.

Added some more command line arguments for testing purposes.